Sécurisation de la route delete post

+++ backend/controllers/post.js
@@ -1,4 +1,5 @@
 const sequelize = require('../middleware/database')
+const jwt = require('jsonwebtoken');
 const Post = require('../models/post')
 const Comment = require('../models/comment')
 const User = require('../models/user')
@@ -42,7 +43,7 @@ exports.findOne = (req, res) => {
 // Update a Posts by the id in the request
 exports.update = (req, res) => {
   const id = req.params.id;
-  Post.update(req.body, { where: { id: id } 
+  Post.update(req.body, { where: { id: id }
   // pour postgresql  returning: true,
   })
     .then((num, post) => {
@@ -57,17 +58,29 @@ exports.update = (req, res) => {
 };
 
 // Delete a Post with the specified id in the request
-exports.delete = (req, res) => {
+exports.delete = async (req, res) => {
   const id = req.params.id;
-  Post.destroy({ where: { id: id } })
-    .then(num => {
-      if (num == 1) {
-        res.status(200).json({ message: "Post was deleted successfully."});
-      } else {
-        res.status(400).json({ message: `Cannot update Post with id=${id}. Maybe Post was not found or req.body is empty!` });
-      }
-    })
-    .catch(error => res.status(500).json({ error }));
+  const token = req.headers.authorization.split(' ')[1];
+  const decodedToken = jwt.verify(token, process.env.SECRET_KEY);
+  const userId = await User.findByPk(decodedToken.userId);
+  const post = await Post.findByPk(id);
+  console.log(req.headers.authorization)
+  console.log(decodedToken.userId)
+  console.log(userId)
+  console.log(post.author)
+  if (userId != post.author) {
+    res.status(403).json({ message: 'Modification non autorisée pour cet utilisateur'});
+  } else {
+    Post.destroy({ where: { id: id } })
+      .then(num => {
+        if (num == 1) {
+          res.status(200).json({ message: "Post was deleted successfully."});
+        } else {
+          res.status(400).json({ message: `Cannot update Post with id=${id}. Maybe Post was not found or req.body is empty!` });
+        }
+      })
+      .catch(error => res.status(500).json({ error }));
+  }
 };
 
 // Comment a Post with the specified id in the request
